refactor: move general-debt to tools
