+++ apps/frontend/lib/tools/index.js
@@ -6,5 +6,6 @@ export {
   default as PrivateStudentLoanForm,
   privateStudentLoanSchema,
 } from "./private-student-loan";
+export { default as GeneralDebtForm, generalDebtSchema } from "./general-debt";
 export * from "./tax-offset";
 export * from "./wage-garnishment";
